requirements.txt
openai
pdfplumber
python-docx
pyyaml
colorama
python-dotenv